Microcode and other things from Xerox PARC archives.
Source: https://xeroxparcarchive.computerhistory.org/indigo/maxc/.index.html

CHM DIGITAL ARCHIVE AGREEMENT

The Computer History Museum (the “Museum”) is willing to provide
access to the Xerox PARC Archive to you only upon the condition that
you accept all the terms contained in this Agreement. By checking on
the “I accept” box in the download form or by viewing, downloading or
otherwise using the Archive, you have indicated that you understand
this Agreement and accept all its terms.

With the permission of the Palo Alto Research Center (PARC), the
Museum is pleased to make available, for non-commercial use only these
Historical Materials.

PARC retains ownership of all intellectual property in and governing
such Historical Materials. Nothing in their use will be construed as a
transfer of ownership interest in the intellectual property. Users
shall not alter or remove any copyright notices and other proprietary
notices on the Historical Materials.

The Historical Materials are being provided “as is” neither PARC nor
the Museum shall have any liability for such Historical Materials and
the delivery of such materials. There is no guarantee of performance
or usability.